+++ crates/tui-realm-stdlib/CHANGELOG.md
@@ -32,6 +32,7 @@
 
 Unreleased
 
+- [Issue 35](https://github.com/veeso/tui-realm-stdlib/issues/35): dont let the `Input`'s cursor escape the component's area
 - Change `ProgressBar` to enable ratatui Gauge option `use_unicode` for smoother progressbars.
 - Fix a bunch of styling inconsistencies. (re #43)